Separate into smaller packages
Introduce dosetup.py script, which is similar to setup.py but accepts
an additional positional option to specify the name of package to process.
eg. python ./dosetup.py ryu-base install --user

The full list of available packages is found in dosetup.py script.

'ryu' is now a meta-package which requires all of the sub-packages.

"python ./setup.py $@" is now an alias of "python ./dosetup.py ryu $@".